+++ .gitignore
@@ -10,3 +10,6 @@ selection-atlas.html
 
 # Jupyter Notebook
 .ipynb_checkpoints
+
+# Python
+*.pyc
